meow

+++ codebook/1_Basic/black_magic.cpp
@@ -8,8 +8,8 @@ int main() {
   heap h1, h2; // max heap
   h1.push(1), h1.push(3), h2.push(2), h2.push(4);
   h1.join(h2); // h1 = {1, 2, 3, 4}, h2 = {};
-  tree<ll, null_type, less<ll>, rb_tree_tag, tree_order_statistics_node_update> st;
-  tree<ll, ll, less<ll>, rb_tree_tag, tree_order_statistics_node_update> mp;
+  tree<int, null_type, less<int>, rb_tree_tag, tree_order_statistics_node_update> st;
+  tree<int, int, less<int>, rb_tree_tag, tree_order_statistics_node_update> mp;
   for (int x : {0, 2, 3, 4}) st.insert(x);
   cout << *st.find_by_order(2) << st.order_of_key(1) << endl; //31
   rope<char> *root[10]; // nsqrt(n)
