+++ codebook/4_Flow_Matching/Kuhn_Munkres.cpp
@@ -1,5 +1,5 @@
-struct KM { // 0-base
-  ll w[N][N], hl[N], hr[N], slk[N];
+struct KM { // 0-base, maximum matching
+  int w[N][N], hl[N], hr[N], slk[N];
   int fl[N], fr[N], pre[N], qu[N], ql, qr, n;
   bool vl[N], vr[N];
   void init(int _n) {
@@ -7,7 +7,7 @@ struct KM { // 0-base
     for (int i = 0; i < n; ++i)
       fill_n(w[i], n, -INF);
   }
-  void add_edge(int a, int b, ll wei) {
+  void add_edge(int a, int b, int wei) {
     w[a][b] = wei;
   }
   bool Check(int x) {
@@ -19,7 +19,7 @@ struct KM { // 0-base
   void bfs(int s) {
     fill_n(slk, n, INF), fill_n(vl, n, 0), fill_n(vr, n, 0);
     ql = qr = 0, qu[qr++] = s, vr[s] = 1;
-    for (ll d;;) {
+    for (int d; ; ) {
       while (ql < qr)
         for (int x = 0, y = qu[ql++]; x < n; ++x)
           if (!vl[x] && slk[x] >= (d = hl[x] + hr[y] - w[x][y])) {
@@ -38,12 +38,12 @@ struct KM { // 0-base
         if (!vl[x] && !slk[x] && !Check(x)) return;
     }
   }
-  ll solve() {
+  int solve() {
     fill_n(fl, n, -1), fill_n(fr, n, -1), fill_n(hr, n, 0);
     for (int i = 0; i < n; ++i)
       hl[i] = *max_element(w[i], w[i] + n);
     for (int i = 0; i < n; ++i) bfs(i);
-    ll res = 0;
+    int res = 0;
     for (int i = 0; i < n; ++i) res += w[i][fl[i]];
     return res;
   }
