+++ codebook/6_Math/Miller_Rabin.cpp
@@ -3,11 +3,11 @@
 // n < 3,474,749,660,383  6 : primes <= 13
 // n < 2^64             7 : 
 // 2, 325, 9375, 28178, 450775, 9780504, 1795265022
-bool Miller_Rabin(ll a, ll n) {
+bool Miller_Rabin(int a, int n) {
   if ((a = a % n) == 0) return 1;
   if (n % 2 == 0) return n == 2;
-  ll tmp = (n - 1) / ((n - 1) & (1 - n));
-  ll t = __lg(((n - 1) & (1 - n))), x = 1;
+  int tmp = (n - 1) / ((n - 1) & (1 - n));
+  int t = __lg(((n - 1) & (1 - n))), x = 1;
   for (; tmp; tmp >>= 1, a = mul(a, a, n))
     if (tmp & 1) x = mul(x, a, n);
   if (x == 1 || x == n - 1) return 1;
