+++ codebook/6_Math/PiCount.cpp
@@ -1,8 +1,8 @@
-ll PrimeCount(ll n) { // n ~ 10^13 => < 2s
+int PrimeCount(int n) { // n ~ 10^13 => < 2s
   if (n <= 1) return 0;
   int v = sqrt(n), s = (v + 1) / 2, pc = 0;
   vector<int> smalls(v + 1), skip(v + 1), roughs(s);
-  vector<ll> larges(s);
+  vector<int> larges(s);
   for (int i = 2; i <= v; ++i) smalls[i] = (i + 1) / 2;
   for (int i = 0; i < s; ++i) {
     roughs[i] = 2 * i + 1;
@@ -19,7 +19,7 @@ ll PrimeCount(ll n) { // n ~ 10^13 => < 2s
       for (int k = 0; k < s; ++k) {
         int i = roughs[k];
         if (skip[i]) continue;
-        ll d = 1LL * i * p;
+        int d = 1LL * i * p;
         larges[ns] = larges[k] - (d <= v ? larges[smalls[d] - pc] : smalls[n / d]) + pc;
         roughs[ns++] = i;
       }
@@ -31,8 +31,8 @@ ll PrimeCount(ll n) { // n ~ 10^13 => < 2s
     }
   }
   for (int k = 1; k < s; ++k) {
-    const ll m = n / roughs[k];
-    ll t = larges[k] - (pc + k - 1);
+    const int m = n / roughs[k];
+    int t = larges[k] - (pc + k - 1);
     for (int l = 1; l < k; ++l) {
       int p = roughs[l];
       if (1LL * p * p > m) break;
