+++ codebook/9_Else/BinarySearchOnFraction.cpp
@@ -1,17 +1,17 @@
 struct Q {
-  ll p, q;
-  Q go(Q b, ll d) { return {p + b.p*d, q + b.q*d}; }
+  int p, q;
+  Q go(Q b, int d) { return {p + b.p*d, q + b.q*d}; }
 };
 bool pred(Q);
 // returns smallest p/q in [lo, hi] such that
 // pred(p/q) is true, and 0 <= p,q <= N
-Q frac_bs(ll N) {
+Q frac_bs(int N) {
   Q lo{0, 1}, hi{1, 0};
   if (pred(lo)) return lo;
   assert(pred(hi));
   bool dir = 1, L = 1, H = 1;
   for (; L || H; dir = !dir) {
-    ll len = 0, step = 1;
+    int len = 0, step = 1;
     for (int t = 0; t < 2 && (t ? step/=2 : step*=2);)
       if (Q mid = hi.go(lo, len + step);
           mid.p > N || mid.q > N || dir ^ pred(mid))
