+++ codebook/9_Else/HilbertCurve.cpp
@@ -1,5 +1,5 @@
-ll hilbert(int n, int x, int y) {
-  ll res = 0;
+int hilbert(int n, int x, int y) {
+  int res = 0;
   for (int s = n / 2; s; s >>= 1) {
     int rx = (x & s) > 0;
     int ry = (y & s) > 0;
